Also rebuilt DLL's, copied to Backend, and updated localizations.

Conflicts:
	Assets/Libs/I2/Localization/Resources/I2Languages.asset
	Assets/Libs/Photon/Quantum/Assemblies/quantum.code.dll
	Assets/Src/FirstLight/Game/StateMachines/MetaAndMatchTutorialState.cs
	Assets/Src/FirstLight/Server/Plugin/net48/FirstLightServerSDK.dll
	Backend/Lib/FirstLight.DataExtensions.dll
	Backend/Lib/FirstLight.Game.Server.dll
	Backend/Lib/FirstLight.Game.dll
	Backend/Lib/FirstLight.Services.dll
	Backend/Lib/PhotonQuantum.dll
	Backend/Lib/quantum.code.dll
	Backend/ServerCommon/Resources/gameConfig.json
	Backend/ServerCommon/Resources/gameTranslations.json
	pipeline-game-auto.yml

It looks like you may be committing a merge.
If this is not correct, please run
	git update-ref -d MERGE_HEAD
and try again.

+++ Assets/Src/FirstLight/Game/Logic/RewardLogic.cs
@@ -174,7 +174,7 @@ namespace FirstLight.Game.Logic
 			// We calculate rank value for rewards based on the actual number of players/teams in a match (including bots)
 			// versus the maximum number of players/teams that are supposed to be in a match. This interpolation is needed
 			// in case we allow rewarded matches with lower number of players, for instance in case we ever do "no bots ranked"
-			var rankValue = Math.Min(1 + Math.Floor(maxTeamsInMatch / (double)((source.GamePlayerCount / teamSize) - 1) * (localMatchData.PlayerRank - 1)), maxTeamsInMatch);
+			var rankValue = (int) Math.Min(1 + Math.Floor(maxTeamsInMatch / (double)((source.GamePlayerCount / teamSize) - 1) * (localMatchData.PlayerRank - 1)), maxTeamsInMatch);
 			
 			foreach (var config in gameModeRewardConfigs)
 			{
